- исправление замечаний после внутреннего код-ревью
