+++ FakeNFT/Docs/Cart.md
@@ -18,7 +18,7 @@
 1. Верстка экрана фильтров (план 2 час, факт 1 час)
 2. Реализация логики фильтрации (план 3 час, факт 3 час)
 3. Реализация экрана удаления NFT из корзины (картинка, вопрос, кнопки удаления и возврата) (план 3 час, факт 5 час)
-4. Реализация сетевого запроса на удаление из корзины NFT (6 час, факт 4 час)
+4. Реализация сетевого запроса на удаление из корзины NFT (план 6 час, факт 4 час)
 Итого время (план): 14 час. 
 Итого время (факт): 13 час.
